♻️ Refactore routes
