+++ lib/routes/documents/eventID.js
@@ -1,9 +1,9 @@
 const { Router } = require("express");
 const { ObjectID } = require("mongodb");
-const { findOne } = require("../api/database");
-const { errorMessages } = require("../api/errorCodes");
-const isValidToken = require("../authentication/isValidToken");
-const { createFunctionSheet } = require("../utils/convertToFunctionSheet");
+const { findOne } = require("../../api/database");
+const { errorMessages } = require("../../api/errorCodes");
+const isValidToken = require("../../authentication/isValidToken");
+const { createFunctionSheet } = require("../../utils/convertToFunctionSheet");
 
 const router = Router();
