+++ lib/routes/documents/overviewDay.js
@@ -1,9 +1,9 @@
 const { Router } = require("express");
-const { find } = require("../api/database");
-const { errorMessages } = require("../api/errorCodes");
-const { buildDailyOverview } = require("../utils/convertToDailyOverview");
-const { createPeriodOfTime } = require("../utils/createPeriodOfTime");
-const isValidToken = require("../authentication/isValidToken");
+const { find } = require("../../api/database");
+const { errorMessages } = require("../../api/errorCodes");
+const { buildDailyOverview } = require("../../utils/convertToDailyOverview");
+const { createPeriodOfTime } = require("../../utils/createPeriodOfTime");
+const isValidToken = require("../../authentication/isValidToken");
 
 const router = Router();
