+++ lib/routes/user.js
@@ -1,7 +1,7 @@
 const { Router } = require("express");
-const login = require("./login");
-const logout = require("./logout");
-const refresh = require("./refresh");
+const login = require("./user/login");
+const logout = require("./user/logout");
+const refresh = require("./user/refresh");
 
 const router = Router();
