+++ lib/routes/user/login.js
@@ -1,9 +1,9 @@
 const jwt = require("jsonwebtoken");
 const { Router } = require("express");
-const { findOne, modifyUser } = require("../api/database");
-const { errorMessages } = require("../api/errorCodes");
-const isValidHeader = require("../authentication/isValidHeader");
-const comparePasswords = require("../authentication/crypto");
+const { findOne, modifyUser } = require("../../api/database");
+const { errorMessages } = require("../../api/errorCodes");
+const isValidHeader = require("../../authentication/isValidHeader");
+const comparePasswords = require("../../authentication/crypto");
 
 const router = Router();
 const EXPIRATION_DATE = 1000 * 60 * 60 * 24;
