+++ lib/routes/user/logout.js
@@ -1,8 +1,8 @@
 const jwt = require("jsonwebtoken");
 const { Router } = require("express");
-const { modifyUser } = require("../api/database");
-const { errorMessages } = require("../api/errorCodes");
-const isValidToken = require("../authentication/isValidToken");
+const { modifyUser } = require("../../api/database");
+const { errorMessages } = require("../../api/errorCodes");
+const isValidToken = require("../../authentication/isValidToken");
 const { ObjectID } = require("mongodb");
 const router = Router();
