+++ lib/routes/user/refresh.js
@@ -1,7 +1,7 @@
 const jwt = require("jsonwebtoken");
 const { Router } = require("express");
-const { findOne } = require("../api/database");
-const { errorMessages } = require("../api/errorCodes");
+const { findOne } = require("../../api/database");
+const { errorMessages } = require("../../api/errorCodes");
 const { ObjectID } = require("mongodb");
 const router = Router();
