+++ server.js
@@ -5,16 +5,14 @@ const cookieParser = require("cookie-parser");
 
 const { connect } = require("./lib/api/database");
 
-const event = require("./lib/routes/eventID");
-const overview = require("./lib/routes/overviewDay");
+const documents = require("./lib/routes/documents");
 const user = require("./lib/routes/user");
 
 const app = express();
 
 app.use(express.json());
 app.use(cookieParser());
-app.use("/api/event", event);
-app.use("/api/date", overview);
+app.use("/api/documents", documents);
 app.use("/api/user", user);
 
 // Heroku Deployment
